Update and rename TripAdvisor2.R to TripAdvisor-Attraction.R
This code is an addendum to the fantastic code Hadley (rvest/demo/tripadvisor.R) put out on Trip Advisor scraping. It is a for-loop specifically for attraction related entities on Trip Advisor, and runs off a CSV of urls for your attraction. 

The html elements are specifically edited to address the way attractions are represented on the Trip Advisor website. If you are working with an attraction that has been around awhile there is a specific code commented out that will get all the dates. Trip Advisor changed the way their dates are coded on the website a couple years ago and that code gets around that.