Create QuickRScraping.R
This is a quick loop that takes a dataframe of two columns, column 1 is the url for the news source. Column 2 is the xml path for the sections you want scraped. Included is the Bing Sentiment value for each article. The resulting dataframe includes the article text, news source line, and sentiment value.